[#84] Move page component to folder pages
