+++ src/Router.jsx
@@ -1,23 +1,23 @@
 import React from "react";
 import { Routes, BrowserRouter, Route } from "react-router-dom";
-import Home from "./components/Home.jsx";
-import Executions from "./components/Executions.jsx";
-import Scripts from "./components/Scripts.jsx";
-import Dagre from "./components/dagre/Dagre.jsx";
-import NoMatch from "./components/NoMatch.jsx";
-import Layout from "./components/Layout.jsx";
+import HomePage from "./pages/HomePage.jsx";
+import ExecutionsPage from "./pages/ExecutionsPage.jsx";
+import ScriptsPage from "./pages/ScriptsPage.jsx";
+import DagrePage from "./pages/dagre/DagrePage.jsx";
+import Layout from "./layouts/Layout.jsx";
+import NotFoundPage from "./pages/NotFoundPage.jsx";
 
 const Router = () => {
   return (
     <BrowserRouter>
       <Routes>
         <Route element={<Layout />}>
-          <Route path="/" element={<Home />} />
-          <Route path="scripts" element={<Scripts />} />
-          <Route path="executions" element={<Executions />} />
-          <Route path="*" element={<NoMatch />} />
+          <Route path="/" element={<HomePage />} />
+          <Route path="scripts" element={<ScriptsPage />} />
+          <Route path="executions" element={<ExecutionsPage />} />
+          <Route path="*" element={<NotFoundPage />} />
         </Route>
-        <Route path="script" element={<Dagre />} />
+        <Route path="script" element={<DagrePage />} />
       </Routes>
     </BrowserRouter>
   );
