+++ src/pages/ExecutionsPage.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import Layout from "./Layout";
 import { ABSOLUTE_PATH, DISPLAY_NAME, EXECUTION_DURATION, TRANSFORMATION } from "../constants/Vocabulary.js";
 import Rest from "../rest/Rest.jsx";
 import { Col, Container, Row, Table } from "react-bootstrap";
@@ -7,7 +6,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faMugHot, faEdit, faQuestion } from "@fortawesome/free-solid-svg-icons";
 import { Link } from "react-router-dom";
 
-class Executions extends React.Component {
+class ExecutionsPage extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -87,4 +86,4 @@ class Executions extends React.Component {
   }
 }
 
-export default Executions;
+export default ExecutionsPage;
