+++ src/pages/HomePage.jsx
@@ -1,14 +1,15 @@
 import React from "react";
 import { Grid } from "semantic-ui-react";
-import Layout from "./Layout";
-import ScriptsTree from "./treebeard/ScriptsTree";
+import ScriptsTree from "../components/treebeard/ScriptsTree.jsx";
 
-class Scripts extends React.Component {
+class HomePage extends React.Component {
   render() {
     return (
       <Grid stackable columns={2}>
         <Grid.Row>
           <Grid.Column>
+            <h1>Welcome to SPipes editor</h1>
+            <br></br>
             <ScriptsTree />
           </Grid.Column>
         </Grid.Row>
@@ -17,4 +18,4 @@ class Scripts extends React.Component {
   }
 }
 
-export default Scripts;
+export default HomePage;
