+++ src/pages/ScriptsPage.jsx
@@ -1,16 +1,13 @@
 import React from "react";
 import { Grid } from "semantic-ui-react";
-import Layout from "./Layout";
-import ScriptsTree from "./treebeard/ScriptsTree";
+import ScriptsTree from "../components/treebeard/ScriptsTree.jsx";
 
-class Home extends React.Component {
+class ScriptsPage extends React.Component {
   render() {
     return (
       <Grid stackable columns={2}>
         <Grid.Row>
           <Grid.Column>
-            <h1>Welcome to SPipes editor</h1>
-            <br></br>
             <ScriptsTree />
           </Grid.Column>
         </Grid.Row>
@@ -19,4 +16,4 @@ class Home extends React.Component {
   }
 }
 
-export default Home;
+export default ScriptsPage;
